chore: create and separate TS interfaces
