+++ frontend/src/interfaces/item.ts
@@ -1,0 +1,30 @@
+export interface Item {
+  _id: string
+  title: string
+  description: string
+  startingBid: number
+  minimumBidIncrement: number
+  endTime: string
+  slug: string
+  sellerId: string
+  currentBid?: number
+  status?: string
+  createdAt?: string
+  updatedAt?: string
+}
+
+export interface CreateItemData {
+  title: string
+  description: string
+  startingBid: number
+  minimumBidIncrement: number
+  endTime: string
+}
+
+export interface UpdateItemData {
+  title?: string
+  description?: string
+  startingBid?: number
+  minimumBidIncrement?: number
+  endTime?: string
+}
